+++ src/main/java/fi/vm/sade/valinta/kooste/viestintapalvelu/route/HyvaksymiskirjeetService.java
@@ -5,6 +5,7 @@ import fi.vm.sade.valinta.kooste.viestintapalvelu.dto.HyvaksymiskirjeDTO;
 import fi.vm.sade.valinta.kooste.viestintapalvelu.dto.KirjeProsessi;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface HyvaksymiskirjeetService {
     void hyvaksymiskirjeetHakemuksille(KirjeProsessi prosessi, HyvaksymiskirjeDTO hyvaksymiskirjeDTO, List<String> hakemusOids);
@@ -14,4 +15,6 @@ public interface HyvaksymiskirjeetService {
     void hyvaksymiskirjeetHakukohteelle(KirjeProsessi prosessi, HyvaksymiskirjeDTO hyvaksymiskirjeDTO);
 
     void hyvaksymiskirjeetHaulle(String hakuOid, String asiointikieli, SijoittelunTulosProsessi prosessi, String defaultValue);
+
+    void hyvaksymiskirjeetHaulleHakukohteittain(String hakuOid, SijoittelunTulosProsessi prosessi, Optional<String> defaultValue);
 }
