+++ src/main/java/fi/vm/sade/valinta/kooste/viestintapalvelu/route/impl/HyvaksymiskirjeetServiceImpl.java
@@ -13,7 +13,9 @@ import fi.vm.sade.sijoittelu.tulos.dto.HakemuksenTila;
 import fi.vm.sade.sijoittelu.tulos.dto.raportointi.HakijaDTO;
 import fi.vm.sade.sijoittelu.tulos.dto.raportointi.HakijaPaginationObject;
 import fi.vm.sade.sijoittelu.tulos.dto.raportointi.HakutoiveenValintatapajonoDTO;
+import fi.vm.sade.tarjonta.service.resources.v1.dto.HakukohdeV1RDTO;
 import fi.vm.sade.valinta.kooste.external.resource.ataru.AtaruAsyncResource;
+import fi.vm.sade.valinta.kooste.external.resource.dokumentti.DokumenttiAsyncResource;
 import fi.vm.sade.valinta.kooste.external.resource.hakuapp.ApplicationAsyncResource;
 import fi.vm.sade.valinta.kooste.external.resource.koodisto.KoodistoCachedAsyncResource;
 import fi.vm.sade.valinta.kooste.external.resource.organisaatio.OrganisaatioAsyncResource;
@@ -34,6 +36,7 @@ import fi.vm.sade.valinta.kooste.viestintapalvelu.dto.Osoite;
 import fi.vm.sade.valinta.kooste.viestintapalvelu.dto.letter.LetterBatch;
 import fi.vm.sade.valinta.kooste.viestintapalvelu.dto.letter.LetterBatchStatusDto;
 import fi.vm.sade.valinta.kooste.viestintapalvelu.dto.letter.LetterResponse;
+import fi.vm.sade.valinta.kooste.viestintapalvelu.dto.letter.TemplateDetail;
 import fi.vm.sade.valinta.kooste.viestintapalvelu.komponentti.HaeOsoiteKomponentti;
 import fi.vm.sade.valinta.kooste.viestintapalvelu.komponentti.HyvaksymiskirjeetKomponentti;
 import fi.vm.sade.valinta.kooste.viestintapalvelu.predicate.SijoittelussaHyvaksyttyHakija;
@@ -67,6 +70,9 @@ import java.util.stream.Collectors;
 @Service
 public class HyvaksymiskirjeetServiceImpl implements HyvaksymiskirjeetService {
     private static final Logger LOG = LoggerFactory.getLogger(HyvaksymiskirjeetServiceImpl.class);
+    private static final String VAKIOTEMPLATE = "default";
+    private static final String VAKIODETAIL = "sisalto";
+
     private final ViestintapalveluAsyncResource viestintapalveluAsyncResource;
     private final HyvaksymiskirjeetKomponentti hyvaksymiskirjeetKomponentti;
     private final ValintaTulosServiceAsyncResource valintaTulosServiceAsyncResource;
@@ -77,6 +83,7 @@ public class HyvaksymiskirjeetServiceImpl implements HyvaksymiskirjeetService {
     private final HaeOsoiteKomponentti haeOsoiteKomponentti;
     private final HakuParametritService hakuParametritService;
     private final KoodistoCachedAsyncResource koodistoCachedAsyncResource;
+    private final DokumenttiAsyncResource dokumenttiAsyncResource;
     private final int pollingIntervalMillis;
 
     private SimpleDateFormat pvmMuoto = new SimpleDateFormat("dd.MM.yyyy");
@@ -94,6 +101,7 @@ public class HyvaksymiskirjeetServiceImpl implements HyvaksymiskirjeetService {
             HaeOsoiteKomponentti haeOsoiteKomponentti,
             HakuParametritService hakuParametritService,
             KoodistoCachedAsyncResource koodistoCachedAsyncResource,
+            DokumenttiAsyncResource dokumenttiAsyncResource,
             @Value("${valintalaskentakoostepalvelu.hyvaksymiskirjeet.polling.interval.millis:10000}") int pollingIntervalMillis) {
         this.viestintapalveluAsyncResource = viestintapalveluAsyncResource;
         this.hyvaksymiskirjeetKomponentti = hyvaksymiskirjeetKomponentti;
@@ -105,6 +113,7 @@ public class HyvaksymiskirjeetServiceImpl implements HyvaksymiskirjeetService {
         this.haeOsoiteKomponentti = haeOsoiteKomponentti;
         this.hakuParametritService = hakuParametritService;
         this.koodistoCachedAsyncResource = koodistoCachedAsyncResource;
+        this.dokumenttiAsyncResource = dokumenttiAsyncResource;
         this.pollingIntervalMillis = pollingIntervalMillis;
     }
 
@@ -359,6 +368,109 @@ public class HyvaksymiskirjeetServiceImpl implements HyvaksymiskirjeetService {
                             prosessi.getPoikkeukset().add(Poikkeus.koostepalvelupoikkeus("Hyväksymiskirjeiden muodostaminen ei onnistunut.\n" + error.getMessage()));
                         }
                 );
+    }
+
+    public void hyvaksymiskirjeetHaulleHakukohteittain(String hakuOid, SijoittelunTulosProsessi prosessi, Optional<String> defaultValue) {
+        valintaTulosServiceAsyncResource.getKoulutuspaikalliset(hakuOid)
+                .flatMap(valintatulokset -> tarjontaAsyncResource.haeHaku(hakuOid)
+                        .flatMap(haku -> {
+                            List<String> hakemusOids = valintatulokset.getResults().stream().map(HakijaDTO::getHakemusOid).collect(Collectors.toList());
+                            return StringUtils.isEmpty(haku.getAtaruLomakeAvain())
+                                    ? applicationAsyncResource.getApplicationsByhakemusOidsInParts(hakuOid, hakemusOids, ApplicationAsyncResource.DEFAULT_KEYS)
+                                    : ataruAsyncResource.getApplicationsByOids(hakemusOids);
+                        })
+                        .flatMap(hakemukset -> {
+                            Map<String, HakemusWrapper> hakemuksetByOid = hakemukset.stream().collect(Collectors.toMap(HakemusWrapper::getOid, h -> h));
+                            Map<String, List<HakijaDTO>> valintatuloksetByHakukohdeOid = valintatulokset.getResults().stream()
+                                    .collect(Collectors.groupingBy(hakija -> hakija.getHakutoiveet().stream()
+                                            .filter(h -> h.getHakutoiveenValintatapajonot().stream().anyMatch(j -> j.getTila().isHyvaksytty()))
+                                            .findAny()
+                                            .get()
+                                            .getHakukohdeOid()));
+                            prosessi.setKokonaistyo(valintatuloksetByHakukohdeOid.size());
+                            return Observable.fromIterable(valintatuloksetByHakukohdeOid.entrySet())
+                                    .flatMap(hakukohteenValintatulokset -> {
+                                        String hakukohdeOid = hakukohteenValintatulokset.getKey();
+                                        List<HakijaDTO> hakukohteenHakijat = hakukohteenValintatulokset.getValue();
+                                        List<HakemusWrapper> hakukohteenHakemukset = hakukohteenHakijat.stream().map(hakija -> hakemuksetByOid.get(hakija.getHakemusOid())).collect(Collectors.toList());
+                                        return tarjontaAsyncResource.haeHakukohde(hakukohdeOid)
+                                                .flatMap(h -> defaultValue.map(Observable::just).orElseGet(() -> haeHakukohteenVakiosisalto(h))
+                                                        .flatMap(vakiosisalto -> {
+                                                            Map<String, MetaHakukohde> hyvaksymiskirjeessaKaytetytHakukohteet = hyvaksymiskirjeetKomponentti.haeKiinnostavatHakukohteet(hakukohteenHakijat);
+                                                            ParametritParser haunParametrit = hakuParametritService.getParametritForHaku(hakuOid);
+                                                            MetaHakukohde kohdeHakukohde = hyvaksymiskirjeessaKaytetytHakukohteet.get(hakukohdeOid);
+                                                            return organisaatioAsyncResource.haeHakutoimisto(kohdeHakukohde.getTarjoajaOid())
+                                                                    .map(toimisto -> ImmutableMap.of(h.getTarjoajaOids().iterator().next(), toimisto.flatMap(t -> Hakijapalvelu.osoite(t, kohdeHakukohde.getHakukohteenKieli()))))
+                                                                    .map(hakijapalveluidenOsoite -> HyvaksymiskirjeetKomponentti.teeHyvaksymiskirjeet(
+                                                                            koodistoCachedAsyncResource::haeKoodisto,
+                                                                            hakijapalveluidenOsoite,
+                                                                            hyvaksymiskirjeessaKaytetytHakukohteet,
+                                                                            hakukohteenHakijat,
+                                                                            hakukohteenHakemukset,
+                                                                            null,
+                                                                            hakuOid,
+                                                                            Optional.empty(),
+                                                                            vakiosisalto,
+                                                                            hakuOid,
+                                                                            "hyvaksymiskirje",
+                                                                            parsePalautusPvm(null, haunParametrit),
+                                                                            parsePalautusAika(null, haunParametrit),
+                                                                            false,
+                                                                            false));
+                                                        }))
+                                                .flatMap(viestintapalveluAsyncResource::viePdfJaOdotaReferenssiObservable)
+                                                .flatMap(letterResponse -> Observable.interval(1, TimeUnit.SECONDS)
+                                                        .flatMap(i -> viestintapalveluAsyncResource.haeStatusObservable(letterResponse.getBatchId())
+                                                                .flatMap(letterBatchStatus -> {
+                                                                    if ("error".equals(letterBatchStatus.getStatus())) {
+                                                                        return Observable.error(new RuntimeException("Viestintäpalvelun statuspyyntö palautti virheen"));
+                                                                    }
+                                                                    if ("ready".equals(letterBatchStatus.getStatus())) {
+                                                                        return Observable.just(letterResponse.getBatchId());
+                                                                    }
+                                                                    return Observable.empty();
+                                                                }))
+                                                        .take(1))
+                                                .flatMap(batchId -> dokumenttiAsyncResource.uudelleenNimea(batchId, "hyvaksymiskirje_" + hakukohdeOid + ".pdf"))
+                                                .map(v -> Pair.of(hakukohdeOid, Optional.<Throwable>empty()))
+                                                .timeout(3, TimeUnit.MINUTES)
+                                                .onErrorReturn(e -> Pair.of(hakukohdeOid, Optional.of(e)));
+                                    }, 6);
+                        }))
+                .subscribe(
+                        result -> {
+                            result.getRight().ifPresentOrElse(e -> {
+                                String msg = String.format("Hakukohteen %s kirjeen muodostaminen epäonnistui", result.getLeft());
+                                LOG.error(msg, e);
+                                prosessi.inkrementoiOhitettujaToita();
+                                prosessi.getPoikkeukset().add(Poikkeus.koostepalvelupoikkeus(msg + "\n" + e.getMessage()));
+                            }, () -> {
+                                LOG.info(String.format("Hakukohteen %s kirjeen muodostaminen onnistui", result.getLeft()));
+                                prosessi.inkrementoi();
+                            });
+                        },
+                        error -> {
+                            String msg = String.format("Haun %s hyväksymiskirjeiden muodostaminen hakukohteittain epäonnistui", hakuOid);
+                            LOG.error(msg, error);
+                            prosessi.getPoikkeukset().add(Poikkeus.koostepalvelupoikkeus(msg + "\n" + error.getMessage()));
+                        },
+                        () -> LOG.info(String.format("Haun %s hyväksymiskirjeiden muodostaminen hakukohteittain onnistui", hakuOid)));
+    }
+
+    private Observable<String> haeHakukohteenVakiosisalto(HakukohdeV1RDTO hakukohde) {
+        return viestintapalveluAsyncResource.haeKirjepohja(
+                hakukohde.getHakuOid(),
+                hakukohde.getTarjoajaOids().iterator().next(),
+                "hyvaksymiskirje",
+                KirjeetHakukohdeCache.getOpetuskieli(hakukohde.getOpetusKielet()),
+                hakukohde.getOid()
+        ).flatMap(kirjepohjat -> kirjepohjat.stream()
+                .filter(kirjepohja -> VAKIOTEMPLATE.equals(kirjepohja.getName()))
+                .flatMap(kirjepohja -> kirjepohja.getTemplateReplacements().stream().filter(tdd -> VAKIODETAIL.equals(tdd.getName())))
+                .map(TemplateDetail::getDefaultValue)
+                .map(Observable::just)
+                .findAny()
+                .orElse(Observable.error(new RuntimeException(String.format("Ei %s tai %s templateDetailia hakukohteelle %s", VAKIOTEMPLATE, VAKIODETAIL, hakukohde.getOid())))));
     }
 
     private List<HakijaDTO> paatteleTarpeellisetKirjeet(List<HakijaDTO> hakijat, List<HakemusWrapper> hakemukset, HyvaksymiskirjeDTO hyvaksymiskirjeDTO) {
